Fixed playlist icon being completely wrong.
The icon was wrong because the enum COggPlayAppUi::TViews was also being used to represent list box item types as well as being used for view types. I've added COggListBox::TItemTypes to represent the types of items that can be displayed in the list box and made some functions return the appropriate type (instead of just returning TInt).