OggPlay.s60.rss OggPlay.s80.rss OggPlay.uiq.rss OggOs.h OggDialogs.cpp: More work on version numbers. They are all in OggOs.h now. Removed the old text strings from the .rss's Updated the non MMF and Motorola version numbers, ready for the next release
OggPlay.cpp:
Fixed playback crackling and stuttering on Motorola phones (tested on A1000 only)
OggPlay's thread priority was too high: it was interfering with the Media server and sound driver causing the stutters. Changed the priority to EThreadPriorityAbsoluteForeground for Motorola builds.

OggTremor_Motorola.h
OggTremor_Motorola.cpp:
Fixed double speed mono playback and fixed stereo too: it was mixing stereo to mono and playing mono as if it was stereo! I also had to add a Resume() fn. to make it compile.