OK I give up. We need to delay the play start until the UI has been updated. Fortunately the streaming thread knows whether it is playing or not.
inc\OggMultiThread.h
src\OggMultiThread.cpp:
PrepareToFlushBuffers() now returns a bool to indicate whether or not streaming has been stopped.
FlushBuffers() now works whether we are streaming or not.

inc\OggTremor.h
src\OggTremor.cpp:
Better handling of timers
Put the restart timer back in (TO DO: transfer restart from the streaming thread to the UI thread)
PrepareToFlushBuffers() returns a TBool (TO DO: Change this, it should be FlushBuffers() that returns the TBool)