Build updates. The MMF vorbis library is now the same as the non-MMF version. All the plugin stuff has been moved into the controller (now a completely seperate dll).
One bug fix. Under some conditions the vorbis code leaks memory. For some reason it was decided that the "fix" for this was to silently exit OggPlay. I've removed this code as silently exiting the application is far worse than leaking a few bytes of memory!

Obviously it would be good to track down the source of the memory leak. It's actually VERY easy to reproduce. Using the non-MMF version if you simply press play, then immediately press pause (i.e. double press the play/pause button) then hit stop you will see the leak. That's actually how I found the problem, I thought OggPlay was crashing!

It's to do with the seek code. What happens is that the playback starts and then pressing pause causes a seek (the MTP engine flushes it's buffers) and then when stop is pressed the memory leak is detected (see the diffs to framing.c for details).

There's nothing obviously wrong, so to track it down will require somebody to step through the vorbis seek function (it too is doing some kind of buffer flush and then buffering up the new data). Why it seems to lose track of some of the buffers I don't know.