File Duration is now reported as part of the file properties Plugin now handles flac as well as vorbis (potentially mp3 too) Correct UIDs for S60V3 (0xFxxxxxxx) Fixed spectrum analyzer (again) Removed most of the debugging (what's the point in release code!?) Removed the sink buffer, it's not at all clear why we create one. Our code certainly doesnt use it. Removing it doesn't seem to make any difference on my E60 (audio dropouts are really bad anyway). Need to check if the N70 playback is still as good or if creating that buffer really makes a difference somehow. The decoders now own the file handle and therefore open and close the audio files themselves(much more sensible).
"Pause" wasn't working on the N70. This was only noticable when playing oggs in the default music player. I tracked this down to iAudioOuput->SinkPauseL(). Calling SinkPauseL() stopped the music but restarting it was then impossible. Changing it to iAudioOutput->SinkStopL() solved the problem. Need to investigate on S60V3 to see if it has the same issue.