Checkin of header files (src to follow)
Changes:
Added flac support
Started work on making the mad plugin work
New dialogs for S60 (and S60V3)

CAbsPlayback::Info() and Open() are now considered to be asynchronous, although the non-MMF versions still complete synchronously as before. The file search code has been modified significantly to handle this. A new macro DELAY_AUDIO_STREAMING_OPEN has been defined in order to handle the MMF's asynchronous file open. I think that maybe I'll remove this in a future release and get the non-MMF version to work in the same way. i.e. NotifyFileOpen() will be called when the file has been opened.

In shuffle mode previous and next track now work as you would expect: the random list of tracks is pre-generated so you can navigate through it.

The view system now correctly tracks if the main view is active or not. The previous code worked in most situations but not always.

The OggPlay version string is now part of the rss (modify the language files to change the version) instead of being in OggOs.h. I think I'll hard code this into the rss (no point having it in every language version!)

Scalable UI support has been improved. Code specifically for scalable UI is now enabled when SERIES60SUI is defined. S60 skins can now be made to support orientation switching. FlipOpen is for portrait orientation, FlipClosed is for landscape.

MDecoder based classes (i.e. the vorbis, flac and mad decoders) now own the file handle and are resposible for opening and closing the file.

Automatic re-scaling of skins has been removed. This wasn't very useful anyway because all of the existing skins try to use the Latin12 font, which isn't scalable. New skins are needed!

The position slider is now an instance of COggSlider64 (a 64bit version of COggSlider)

The S60 settings views are no longer one control inside another (that was a bit pointless)

Some of the truly ancient macros have been removed (i.e. playlist support is no longer optional, the spectrum analyzer is the MDCT variant, multi-thread playback is now the only non-MMF playback engine etc.). I've also started getting rid of MOTOROLA, I will get rid of the rest of them when I get around to it.

White space is now mostly compliant with Symbian coding standards (tabs not spaces, braces in-line with the code).

I think that's about it. Oh yeah, one more thing. I fiddled with some of the enums in OggPlay.hrh,  we were using the wrong ones for some of the settings controls