Added new S60V3 readme files (removed the old ones)
Added mad\madplay folder
These are the files we've pulled in from the madplay project
(they were previously in the src folder, but it makes far more sense to have them in a folder called madplay, making it much clearer where they came from)

The rgainx.h and xltagx.h  headers are identical to rgain.h and xltag.h, except for the "includes".
It proved to be tricky to get rgain.h and xltag.h to work correctly when included in both libmad and in oggplay, so I just duplicated the files and changed the includes. It's not an ideal solution, but  it will do.