OggPlayController.cpp: mp3 support in the plugin requires a different macro to be defined
OggPlayController*.mmp:
Added MMF_PLUGIN macro
This is only needed to avoid including the jcode stuff (supposedly it doesn't work in the MMF plugin, but I suspect this may no longer be the case). Ideally, the issues with jcode and MMF should be investigated and this macro removed if it is no longer needed.

Release notes:
For S60 only:
Updated to 1.72

Install files:
For S60V2 only:
Put the libs in a seperate package. This is embedded in the main sis.
Added MMF plugin package (also embedded the libs sis)