FlacDecoder.cpp: Check that iBuf gets allocated >>1 is faster than / 2, right. Presumably the compiler should do this sort of thing for you anyway...
FlacDecoder.h:
Different size buffers
The issue isn't with memory, it's the time taken to copy the data. Because of the double buffering the streaming thread must wait for the current read to complete before transfering the buffering to the buffering thread. This could easily cause the stream to stall and is the reason for using smaller buffers on older hardware (i.e. the Sendo)

What's needed is to change the streaming thread into an IO thread. The streaming thread should do the file reading as well as the audio streaming. That would avoid this problem and give us the maximum possible performance (allowing as much overlapping of decoding and file reading as possible).